chore: Moved table in test component to its own component. There is still an underlying issue where backend refuses to do anything anymore. I'm still not sure what causes this, but it's no longer happening.
