+++ src/app/test/test-table/test-table.component.css
@@ -1,3 +1,7 @@
+:host {
+  display: content !important;
+}
+
 td {
   border: 0;
   vertical-align: middle;
