+++ src/app/test/test-table/test-table.component.html
@@ -1,0 +1,75 @@
+<div class="table-responsive">
+  <table class="table table-hover">
+    <thead>
+    <tr>
+      <th>
+        <input data-cy-test="toggleSelectAll" type="checkbox" [ngModel]="amountOfSelectedReports === reports?.length && amountOfSelectedReports > 0"
+               (click)="toggleSelectAll()">
+      </th>
+      @if (showStorageIds) {
+        <th>StorageId</th>
+      }
+      <th>(Re)run/open</th>
+      <th>Name</th>
+      <th>Description</th>
+      <th>Variables</th>
+      <th>Run result</th>
+      <th>Options</th>
+    </tr>
+    </thead>
+    <tbody data-cy-test="table" id="testReports">
+      @for (report of reports; let index = $index; track report.storageId) {
+        @if (showReport[index]) {
+          <tr>
+            <td>
+              <input data-cy-test="reportChecked" type="checkbox" [checked]="report.checked"
+                     (click)="toggleCheck(report)">
+            </td>
+            @if (showStorageIds) {
+              <td>
+                {{ report.storageId }}
+              </td>
+            }
+            <td>
+              <span class="buttons">
+                <button data-cy-test="runReport" title="Run report" class="btn btn-info button-width"
+                        (click)="run(report)"><i class="fa fa-play"></i></button>
+                <button data-cy-test="openReport" title="Open report" class="btn btn-info button-width"
+                        (click)="openReport(report.storageId)"><i class="fa fa-folder-open"></i>
+                </button>
+              </span>
+            </td>
+            <td>
+              {{ getFullPath(report.path, report.name) }}
+            </td>
+            <td>
+              {{ report.description }}
+            </td>
+            <td>
+              @if (report.variables) {
+                <span data-cy-test="variables">{{ extractVariables(report.variables) }}</span>
+              }
+            </td>
+            @if (report.reranReport; as reranReport) {
+              <td>
+                  <span data-cy-test="runResult"
+                        [id]="'runResult#' + report.storageId"
+                        [style.color]="reranReport.color">
+                              {{ reranReport.resultString }}
+                  </span>
+              </td>
+              <td>
+                  <span class="buttons">
+                    <button data-cy-test="replaceReport" class="btn btn-info" title="Replace report"
+                            (click)="replaceReport(report.storageId)">Replace</button>
+                    <button data-cy-test="compareReport" class="btn btn-info" title="Compare"
+                            (click)="compareReports(report)">Compare</button>
+                  </span>
+              </td>
+            }
+          </tr>
+        }
+      }
+    </tbody>
+  </table>
+</div>
