+++ src/app/test/test-table/test-table.component.spec.ts
@@ -1,20 +1,20 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { TestTableBodyComponent } from './test-table-body.component';
-import { HttpClient, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
+import { TestTableComponent } from './test-table.component';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { provideHttpClientTesting } from '@angular/common/http/testing';
 
 describe('TestTableBodyComponent', () => {
-  let component: TestTableBodyComponent;
-  let fixture: ComponentFixture<TestTableBodyComponent>;
+  let component: TestTableComponent;
+  let fixture: ComponentFixture<TestTableComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [TestTableBodyComponent],
+      imports: [TestTableComponent],
       providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()],
     }).compileComponents();
 
-    fixture = TestBed.createComponent(TestTableBodyComponent);
+    fixture = TestBed.createComponent(TestTableComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
